Refactor move datatypes/schemas.py to managers/datatypes.py
